Move 404 page to app router
