+++ frontend/src/app/not-found.tsx
@@ -1,12 +1,23 @@
-import BetaAlert from "src/components/AppBetaAlert";
+import BetaAlert from "src/components/BetaAlert";
 import { GridContainer } from "@trussworks/react-uswds";
 import Link from "next/link";
 import { useTranslations } from "next-intl";
-import { unstable_setRequestLocale } from "next-intl/server";
+import { getTranslations, unstable_setRequestLocale } from "next-intl/server";
+import { Metadata } from "next";
+
+export async function generateMetadata() {
+  const t = await getTranslations({ locale: "en" });
+  const meta: Metadata = {
+    title: t("ErrorPages.page_not_found.title"),
+    description: t("Index.meta_description"),
+  };
+  return meta;
+}
 
 export default function NotFound() {
   unstable_setRequestLocale("en");
   const t = useTranslations("ErrorPages.page_not_found");
+
   return (
     <>
       <BetaAlert />
